+++ docs/source/changelog.rst
@@ -9,7 +9,11 @@ HEAD (TBR)
 **********
 
 * Fix traceback when a certificate that does not exist is viewed in the admin interface.
-* Support cryptography 2.5 and 2.6.
+* Add support cryptography 2.5 and 2.6.
+* New settings :ref:`CA_FILE_STORAGE <settings-ca-file-storage>` and :ref:`CA_FILE_STORAGE_KWARGS
+  <settings-ca-file-storage-kwargs>` to support a custom `Django storage backend
+  <https://docs.djangoproject.com/en/2.1/ref/files/storage/>`_ (e.g. one from `django-storages
+  <https://django-storages.readthedocs.io/en/latest/>`_). 
 * **BACKWARDS INCOMPATIBLE:** Drop support for Python 3.4.
 * **BACKWARDS INCOMPATIBLE:** Drop support for Django 2.0.
 * **BACKWARDS INCOMPATIBLE:** Drop support for cryptography 2.1.
@@ -20,6 +24,10 @@ OCSP
 ====
 
 * Reimplement OCSP using cryptography, used only if cryptography>=2.4 is installed.
+* :py:class:`django_ca.views.OCSPBaseView.responder_key` may now also be a relative path to be used with the 
+  Django storage system.
+* :py:class:`django_ca.views.OCSPBaseView.responder_cert` may now also be a relative path to be used with the 
+  Django storage system.
 * :py:class:`django_ca.views.OCSPBaseView.responder_cert` may now also be a pre-loaded certificate. If you
   still use ``cryptography<2.4`` use a ``oscrypto.asymmetric.Certificate``, for newer versions you must
   use a :py:class:`cg:cryptography.x509.Certificate`.